backend/requirements.txt
# --- Core ---
fastapi==0.115.2
uvicorn==0.31.1
strawberry-graphql==0.242.0  # GraphQL framework basado en FastAPI

# --- Base de datos (PostgreSQL) ---
sqlalchemy==2.0.34
psycopg2
asyncpg==0.29.0              # Driver asíncrono para PostgreSQL
alembic==1.13.3              # Migraciones de base de datos

# --- Validación y configuración ---
pydantic==2.8.2
pydantic-settings==2.5.2
python-dotenv==1.0.1         # Manejo de variables .env

# --- Logging y utilidades ---
loguru==0.7.2

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2                # Pruebas de endpoints async
